Introduce a simulation context
Adds `SimulationContext` class which encapsulates information about shipments of a simulation instance. The `formRecordSet` method groups the raw record data from the XML file into `Shipment`s. A `Shipment` represents an incoming or outgoing shipment by a carrier which carries a certain amount of containers and is is identified by a key based on the arrival or departure data. The shipment data can be accessed by the `getShipments` method. Shipments by a specific carrier type can be access with `getX` where `X` is either `SeaShips`, `InlandShips`, `Trucks` or `Trains`.